pix-644 issue-137 - move shared folder
